Updated imports, renamed methods, and fixed a bug in the 'view' layer.
