+++ CoworkingSpaceReservation/src/main/java/org/example/controller/CustomerController.java
@@ -1,8 +1,8 @@
 package org.example.controller;
 
-import org.example.model.dto.ReservationDto;
-import org.example.model.dto.UserDto;
-import org.example.model.dto.WorkspaceDto;
+import org.example.model.dto.space.ReservationDto;
+import org.example.model.dto.account.UserDto;
+import org.example.model.dto.space.WorkspaceDto;
 import org.example.model.service.CustomerManager;
 import org.example.view.CustomerView;
