+++ CoworkingSpaceReservation/src/main/java/org/example/controller/factory/AdminFactory.java
@@ -1,13 +1,13 @@
 package org.example.controller.factory;
 
 
-import org.example.model.dto.UserDto;
+import org.example.model.dto.account.UserDto;
 import org.example.model.service.AdminManager;
 
 class AdminFactory implements UserFactory {
     @Override
     public UserDto createUser() {
-        return AdminManager.getInstance().buildAdmin();
+        return AdminManager.getInstance().getEmptyAdmin();
     }
 }
