+++ CoworkingSpaceReservation/src/main/java/org/example/controller/factory/CustomerFactory.java
@@ -1,11 +1,11 @@
 package org.example.controller.factory;
 
-import org.example.model.dto.UserDto;
+import org.example.model.dto.account.UserDto;
 import org.example.model.service.CustomerManager;
 
 class CustomerFactory implements UserFactory {
     @Override
     public UserDto createUser() {
-        return CustomerManager.getInstance().buildCustomer();
+        return CustomerManager.getInstance().getEmptyCustomer();
     }
 }
