+++ CoworkingSpaceReservation/src/main/java/org/example/controller/navigation/command/customer/CustomerCommand.java
@@ -2,7 +2,7 @@ package org.example.controller.navigation.command.customer;
 
 import lombok.Getter;
 import org.example.controller.navigation.command.MenuCommand;
-import org.example.model.dto.CustomerDto;
+import org.example.model.dto.account.CustomerDto;
 @Getter
 public abstract class CustomerCommand implements MenuCommand {
     private final CustomerDto customer;
