+++ CoworkingSpaceReservation/src/main/java/org/example/model/service/ReservationManager.java
@@ -3,18 +3,21 @@ package org.example.model.service;
 import org.example.exceptions.ReservationAlreadyExistException;
 import org.example.exceptions.ReservationNotFoundForUserException;
 import org.example.exceptions.WorkspaceAlreadyBookedAtTimeException;
-import org.example.model.dao.ReservationDao;
-import org.example.model.dao.WorkspaceDao;
-import org.example.model.dto.CustomerDto;
-import org.example.model.dto.ReservationDto;
-import org.example.model.dto.WorkspaceDto;
-import org.example.model.dto.filters.Filter;
-import org.example.model.dto.filters.ReservationFilter;
-import org.example.model.entity.CustomerEntity;
-import org.example.model.entity.ReservationEntity;
-import org.example.model.entity.WorkspaceEntity;
+import org.example.model.dto.account.CustomerDto;
+import org.example.model.dto.space.ReservationDto;
+import org.example.model.dto.space.WorkspaceDto;
+import org.example.model.dto.filters.space.ReservationFilter;
+import org.example.model.entity.account.CustomerEntity;
+import org.example.model.entity.space.ReservationEntity;
+import org.example.model.entity.space.SpaceTypeEntity;
+import org.example.model.entity.space.WorkspaceEntity;
+import org.example.model.repository.space.ReservationRepository;
+import org.example.model.repository.space.WorkspaceRepository;
+import org.example.model.util.SessionManager;
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
 
-import java.sql.SQLException;
 import java.time.LocalDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
